+++ packages/online-editor/src/dmnRunnerInputs/DmnRunnerInputsHook.tsx
@@ -17,9 +17,10 @@
 import * as React from "react";
 import { useCallback, useRef, useState } from "react";
 import { useCancelableEffect, usePreviousRef } from "../reactExt/Hooks";
-import { decoder, WorkspaceFile } from "../workspace/WorkspacesContext";
+import { WorkspaceFile } from "../workspace/WorkspacesContext";
 import { InputRow } from "@kie-tools/form-dmn";
 import { useDmnRunnerInputsDispatch } from "./DmnRunnerInputsContext";
+import { decoder } from "../workspace/encoderdecoder/EncoderDecoder";
 
 interface DmnRunnerInputs {
   inputRows: Array<InputRow>;
