+++ packages/online-editor/src/editor/FileSwitcher.tsx
@@ -15,7 +15,7 @@
  */
 
 import { ActiveWorkspace } from "../workspace/model/ActiveWorkspace";
-import { decoder, useWorkspaces, WorkspaceFile } from "../workspace/WorkspacesContext";
+import { useWorkspaces, WorkspaceFile } from "../workspace/WorkspacesContext";
 import { useRoutes } from "../navigation/Hooks";
 import * as React from "react";
 import { useCallback, useEffect, useMemo, useRef, useState } from "react";
