+++ packages/online-editor/src/workspace/hooks/WorkspaceHooks.tsx
@@ -21,6 +21,7 @@ import { usePromiseState } from "./PromiseState";
 import { Holder, useCancelableEffect } from "../../reactExt/Hooks";
 import { WorkspaceKind } from "../worker/api/WorkspaceOrigin";
 import { GIT_ORIGIN_REMOTE_NAME } from "../constants/GitConstants";
+import { WorkspaceEvents } from "../worker/api/WorkspaceEvents";
 
 export function useWorkspaceGitStatusPromise(workspace: ActiveWorkspace | undefined) {
   const workspaces = useWorkspaces();
@@ -151,18 +152,3 @@ export function useWorkspacePromise(workspaceId: string | undefined) {
 
   return workspacePromise;
 }
-
-export type WorkspaceEvents =
-  | { type: "ADD"; workspaceId: string }
-  | { type: "CREATE_SAVE_POINT"; workspaceId: string }
-  | { type: "PULL"; workspaceId: string }
-  | { type: "RENAME"; workspaceId: string }
-  | { type: "DELETE"; workspaceId: string }
-  | { type: "ADD_FILE"; relativePath: string }
-  | { type: "MOVE_FILE"; newRelativePath: string; oldRelativePath: string }
-  | { type: "RENAME_FILE"; newRelativePath: string; oldRelativePath: string }
-  | { type: "UPDATE_FILE"; relativePath: string }
-  | { type: "DELETE_FILE"; relativePath: string }
-  | { type: "ADD_BATCH"; workspaceId: string; relativePaths: string[] }
-  | { type: "MOVE_BATCH"; workspaceId: string; relativePaths: Map<string, string> }
-  | { type: "DELETE_BATCH"; workspaceId: string; relativePaths: string[] };
